Moved the MoveHistory into a separate class.
